Fix: results table html hook
